+++ branches/x.3/examples/10.Libraries/EEPROM/eeprom_write.pde
@@ -10,7 +10,7 @@
 
 // the current address in the EEPROM (i.e. which byte
 // we're going to write to next)
-int addr = 0;
+u16 addr = 0;
 
 void setup()
 {
